requirements.txt
pypdf>=3.17.1
langchain>=0.0.267
langchain_community>=0.0.10
nltk>=3.8.1
python-dotenv>=1.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
